[tools] selftests: bpf: use __u32 for sock::sk_pacing_rate
Message-id: <20200713134744.145676-1-yauheni.kaliuta@redhat.com>
Patchwork-id: 322148
Patchwork-instance: patchwork
O-Subject: [PATCH RHEL8.3] selftests: bpf: use __u32 for sock::sk_pacing_rate
Bugzilla: 1856324
RH-Acked-by: Jiri Benc <jbenc@redhat.com>
RH-Acked-by: Vladis Dronov <vdronov@redhat.com>

Upstream: RHEL-only.
Bugzilla: https://bugzilla.redhat.com/1856324
Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=30075024

Due to missing backport of

76a9ebe811fb ("net: extend sk_pacing_rate to unsigned long")

the test performs access of wrong size.

Sync test/userspace and kernel header in this regard.

+++ tools/testing/selftests/bpf/bpf_tcp_helpers.h
@@ -26,7 +26,7 @@ enum sk_pacing {
 
 struct sock {
 	struct sock_common	__sk_common;
-	unsigned long		sk_pacing_rate;
+	__u32			sk_pacing_rate;
 	__u32			sk_pacing_status; /* see enum sk_pacing */
 } __attribute__((preserve_access_index));
